Fixed a bug in the single-pass stats routine while folding.
Thanks to nagging by Paul Ray I found this bug.  It was a simple
off-by-one error in the one-pass statistics algorithm which
resulted in the data variance being too high.  For well-behaved
(i.e. stationary gaussian noise) data, that caused a bit hit in
chi^2 in the first sub-fold (i.e. "part") in time.  This _might_
be the cause of the large number of bogus negative-sigma candidate
folds in the surveys.  However, it doesn't fix problems with
impulsive RFI...for that you almost certainly need to fold using
-runavg.