Made the -runavg slightly more robust.
It now subtracts a running average of the block averages rather
than the current block average itself.  That makes it much safer
when folding real pulsars.  Note that there is still a
"conservation of flux" here, so you will see dips before and after
real pulses if you fold with this!  It does help a lot at removing
impulsive RFI and changing system noise/gain, though.