Revert "Permit multiple input files on the command line in order to avoid"
This conflicts with the clig command line parser / doc generator files,
and also needs further optimization on command line parsing.

This also missed a call to system(3).
Leaving it there partly defeated the purpose of this commit,
which is to reduce the number of shell executions and associated
searched over $PATH and $LD_LIBRARY_PATH.

This reverts commit f0256338b54f833962d8c39a7facf4a941043c05.